APP-17935 Restructure to allow cloud ds tests to run on Local
