+++ dsemulator_test.go
@@ -1,0 +1,7 @@
+// +build !cdsemulator
+
+package appwrap
+
+func (s *AppengineInterfacesTest) newDatastore() Datastore {
+	return NewLocalDatastore(false, nil)
+}
